Add error calculation measurement code
Add loops performing multiple multiply-and-add, multiply and addition matrix
calculations to assess the level of accumulated error introduced over multiple
applications.

The code outputs and reads in CSV results to combine into graphs, so that
results from difference systems can be plotted together.

Includes scripts for running the tests on Baskerville and DAWN.