Update Baskerville sbatch and srun scripts
Updates the scripts for queuing using sbatch or running within an interactive
shell on a compute node using srun.

The objective with these changes is to ensure:

1. Everything works simply by running the scripts.
2. There's an element of consistency across the various scripts.

The minimum xarray version has been increased on Baskerville to 2023.06.0 in
order to avoid Issue #7880 which I experienced consistently when executing the
runmodel.py version. See here for the relevant xarray issue and changelogs:

https://github.com/pydata/xarray/pull/7880

https://github.com/pydata/xarray/releases/tag/v2023.06.0