Add draft tech report document
Adds a draft of the technical report explaining our findings running
Aurora on Dawn and comparing it with Baskerville.

Also adds a LaTeX build pipeline for automatically building the PDF from
the LaTeX source.

Adds a badge to the root-level README to indicate the build result and
allow the resulting PDF to be easily downloaded.